little fix types

+++ configs/init.go
@@ -11,8 +11,8 @@ type Configuration struct {
 	Token             string `json:"token"`
 	LogFile           string `json:"logFile"`
 	UnwrittenDataFile string `json:"unwrittenDataFile"`
-	AccessChatID      int    `json:"accessChatID"`
-	AdminTgChatID     int    `json:"adminTgChatID"`
+	AccessChatID      int64  `json:"accessChatID"`
+	AdminTgChatID     int64  `json:"adminTgChatID"`
 	LinkToDB          string `json:"linkToDB"`
 	Queries           struct {
 		AddUser     string `json:"addUser"`
